pyproject.toml
[tool.poetry]
name = "cdk-webapp-skeleton"
version = "0.3.dev15"
description = ""
authors = ["Ilya Nekhay <nekhayiv@gmail.com>"]
readme = "README.md"
packages = [{include = "cdk_webapp_skeleton"}]

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = "^2.67.0"
cloudcomponents-cdk-static-website = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pre-commit = "^3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=. --cov-report html --cov-config=.coveragerc"
markers = [
    "branch_name",
]

[tool.isort]
profile = "black"
multi_line_output = 3
